:wrench: 调整引入 Spring 的方式，不再直接依赖插件；已测试对 SpringBoot1.x 的支持

+++ src/test/java/org/cat73/pager/aop/PagerPluginTest.java
@@ -159,8 +159,8 @@ class PagerPluginTest extends SpringBootTestBase {
             while (rowIterator.hasNext()) {
                 row = rowIterator.next();
 
-                Assertions.assertEquals(i, row.getCell(0).getNumericCellValue());
-                Assertions.assertEquals(n, row.getCell(1).getNumericCellValue());
+                Assertions.assertEquals((double) i, row.getCell(0).getNumericCellValue());
+                Assertions.assertEquals((double) n, row.getCell(1).getNumericCellValue());
                 Assertions.assertEquals("node-" + n, row.getCell(2).getStringCellValue());
 
                 i++;
@@ -201,8 +201,8 @@ class PagerPluginTest extends SpringBootTestBase {
             while (rowIterator.hasNext()) {
                 row = rowIterator.next();
 
-                Assertions.assertEquals(i, row.getCell(0).getNumericCellValue());
-                Assertions.assertEquals(n, row.getCell(1).getNumericCellValue());
+                Assertions.assertEquals((double) i, row.getCell(0).getNumericCellValue());
+                Assertions.assertEquals((double) n, row.getCell(1).getNumericCellValue());
                 Assertions.assertEquals("node-" + n, row.getCell(2).getStringCellValue());
 
                 i++;
